+++ cmd/routed-eni-cni-plugin/cni.go
@@ -46,7 +46,7 @@ import (
 )
 
 const (
-	ipamdAddress            = "127.0.0.1:50051"
+	ipamdSocketPath         = "/var/run/aws-node/ipamd.sock"
 	dummyInterfacePrefix    = "dummy"
 	npAgentConnTimeout      = 2
 	npaSocketPath           = "/var/run/aws-node/npa.sock"
@@ -136,6 +136,15 @@ func LoadNetConf(bytes []byte) (*NetConf, logger.Logger, error) {
 	return &conf, log, nil
 }
 
+func dialIPAMD(grpcClient grpcwrapper.GRPC, log logger.Logger, socketPath string) (*grpc.ClientConn, error) {
+	conn, err := grpcClient.Dial("unix://"+socketPath, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	if err != nil {
+		return nil, err
+	}
+	log.Debugf("Connected to IPAMD via Unix socket: %s", socketPath)
+	return conn, nil
+}
+
 func cmdAdd(args *skel.CmdArgs) error {
 	return add(args, typeswrapper.New(), grpcwrapper.New(), rpcwrapper.New(), driver.New())
 }
@@ -177,8 +186,8 @@ func add(args *skel.CmdArgs, cniTypes typeswrapper.CNITYPES, grpcClient grpcwrap
 
 	log.Debugf("pod requires multi-nic attachment: %t", requiresMultiNICAttachment)
 
-	// Set up a connection to the ipamD server.
-	conn, err := grpcClient.Dial(ipamdAddress, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	// Set up a connection to the ipamD server via Unix socket.
+	conn, err := dialIPAMD(grpcClient, log, ipamdSocketPath)
 	if err != nil {
 		log.Errorf("Failed to connect to backend server for container %s: %v",
 			args.ContainerID, err)
@@ -365,11 +374,10 @@ func del(args *skel.CmdArgs, cniTypes typeswrapper.CNITYPES, grpcClient grpcwrap
 	driverClient driver.NetworkAPIs) error {
 
 	conf, log, err := LoadNetConf(args.StdinData)
-	log.Debugf("Prev Result: %v\n", conf.PrevResult)
-
 	if err != nil {
 		return errors.Wrap(err, "del cmd: error loading config from args")
 	}
+	log.Debugf("Prev Result: %v\n", conf.PrevResult)
 
 	log.Infof("Received CNI del request: ContainerID(%s) Netns(%s) IfName(%s) Args(%s) Path(%s) argsStdinData(%s)",
 		args.ContainerID, args.Netns, args.IfName, args.Args, args.Path, args.StdinData)
@@ -395,8 +403,8 @@ func del(args *skel.CmdArgs, cniTypes typeswrapper.CNITYPES, grpcClient grpcwrap
 	}
 
 	// notify local IP address manager to free secondary IP
-	// Set up a connection to the server.
-	conn, err := grpcClient.Dial(ipamdAddress, grpc.WithInsecure())
+	// Set up a connection to the server via Unix socket.
+	conn, err := dialIPAMD(grpcClient, log, ipamdSocketPath)
 	if err != nil {
 		log.Errorf("Failed to connect to backend server for container %s: %v",
 			args.ContainerID, err)